+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to [bpmn-js-properties-panel](https://github.com/bpmn-io/bpm
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: add `propertiesPanel#getEntryId(element, path)` API to resolve a moddle property path to an entry id
 * `FEAT`: use `number` input for `priority` and `jobPriority` ([#1241](https://github.com/bpmn-io/bpmn-js-properties-panel/pull/1241))
 
 ## 5.62.0